Update AppliedJob API

+++ src/main/java/app/service/AppliedJobService.java
@@ -6,6 +6,8 @@ import javax.persistence.EntityNotFoundException;
 
 import org.springframework.stereotype.Service;
 import app.entity.AppliedJob;
+import app.entity.Member;
+import app.exception.UnauthorizedException;
 import app.repository.AppliedJobRepository;
 
 @Service
@@ -16,24 +18,42 @@ public class AppliedJobService
 	{
 		this.appliedJobRepository=appliedJobRepository;
 	}
+	public boolean existsById(Long id)
+	{
+		if(appliedJobRepository.existsById(id))
+			return true;
+		throw new EntityNotFoundException();
+	}
+	public boolean belongsTo(Long id,Member member)
+	{
+		AppliedJob appliedJob=appliedJobRepository.findById(id).get();
+		if(appliedJob.getMember().equals(member))
+			return true;
+		throw new UnauthorizedException();
+	}
 	public AppliedJob save(AppliedJob appliedJob)
 	{
 		return appliedJobRepository.save(appliedJob);
 	}
-	public AppliedJob update(AppliedJob appliedJob)
+	public void deleteById(Long id,Member member)
 	{
-		return appliedJobRepository.save(appliedJob);
+		if(existsById(id))
+		{
+			if(belongsTo(id,member))
+				appliedJobRepository.deleteById(id);
+		}
 	}
-	public void deleteById(Long id)
+	public AppliedJob findById(Long id,Member member)
 	{
-		appliedJobRepository.deleteById(id);
+		if(existsById(id))
+		{
+			if(belongsTo(id,member))
+				return appliedJobRepository.findById(id).get();
+		}
+		return null;
 	}
-	public AppliedJob findById(Long id)
+	public List<AppliedJob> findByMember(Member member)
 	{
-		return appliedJobRepository.findById(id).orElseThrow(EntityNotFoundException::new);
-	}
-	public List<AppliedJob> findAll()
-	{
-		return appliedJobRepository.findAll();
+		return appliedJobRepository.findByMember(member);
 	}
 }
